fix: send again toolkit/init image data along with Pod information from Test Workflow
